+++ resolwe_bio/tests/workflows/test_wgs.py
@@ -13,14 +13,15 @@ class WgsWorkflowTestCase(BioProcessTestCase):
             return line.startswith(b"##GATKCommandLine")
 
         with self.preparation_stage():
-            genome = self.run_process(
-                "upload-genome",
+            ref_seq = self.run_process(
+                "upload-fasta-nucl",
                 {
                     "src": "./bqsr/input/hs_b37_chr17_upto_TP53.fasta.gz",
                     "species": "Homo sapiens",
                     "build": "custom_build",
                 },
             )
+            bwa_index = self.run_process("bwa-index", {"ref_seq": ref_seq.id})
 
             reads = self.prepare_paired_reads(
                 mate1=["./workflow_wes/input/TP53_1.fastq.gz"],
@@ -40,7 +41,8 @@ class WgsWorkflowTestCase(BioProcessTestCase):
         self.run_process(
             "workflow-wgs-paired",
             {
-                "genome": genome.id,
+                "bwa_index": bwa_index.id,
+                "ref_seq": ref_seq.id,
                 "reads": reads.id,
                 "known_sites": [i.id for i in kbase],
                 "hc_dbsnp": kbase[0].id,